requirements.txt
requests
python-dotenv
pyperclip
